MungaBot v6.2.0
__New Commands:__
`>coinflip`
`>info-coinflip`
__Commands Updated:__
`>info`
`>dates`
`>instagram`
`>poll`
`>spoiler`
`>version`
`>userinfo`
`>weather`
__Notes:__
`>info-admin` and `>info-media` and `>info-utility` merged to `>info-commands`, which only leaves `>info-commands`, `>info-cards` and `>info-socialmedia`

+++ commands/spoiler.js
@@ -3,7 +3,18 @@ module.exports = {
     description: "spoiler",
     execute(message, args){
         let letter = args.join("");
-        message.delete();
-        message.channel.send("||" + letter.split("").join('||||') + "||");
+        const Discord = require('discord.js')
+
+        if(!args[0]){
+            const embedHELP = new Discord.MessageEmbed()
+            .setColor(0xFF0000)
+            .addField("Correct usage:", "`>spoiler` <...>")
+            .setFooter("For more info use `>info-spoiler`");
+            message.channel.send(embedHELP);
+        } else {
+            message.delete();
+            message.channel.send("||" + letter.split("").join('||||') + "||");
+        }
+
     }
 }
